Getting tired - time to commit before I make a silly. Remember to do more error checking.
Changed a lot this time, completely redid the colour scheme. Still troubled with the psd for the banner, cannot remember, for the life of me, how to make it transparent again.
Reformatted a lot of different items, and fiddled with colours. Started adding big divs. Remember to make the goal of each page visible.
Commit done.